+++ src/source/ATech.Ring.DotNet.Cli/Abstractions/Runnable.cs
@@ -38,8 +38,9 @@ namespace ATech.Ring.DotNet.Cli.Abstractions
         /// <param name="value"></param>
         protected void AddDetail(string key, object value) => _details.TryAdd(key, value);
 
-        protected Runnable(ILogger<Runnable<TContext, TConfig>> logger, ISender<IRingEvent> sender)
+        protected Runnable(TConfig config, ILogger<Runnable<TContext, TConfig>> logger, ISender<IRingEvent> sender)
         {
+            Config = config;
             _logger = logger;
             Sender = sender;
         }
@@ -158,9 +159,8 @@ namespace ATech.Ring.DotNet.Cli.Abstractions
             }
         }
 
-        public async Task RunAsync(IRunnableConfig config, CancellationToken token)
+        public async Task RunAsync(CancellationToken token)
         {
-            Config = (TConfig)config;
             using var _ = _logger.BeginScope(this.ToScope());
             var fsm = await InitFsm(token);
 
@@ -203,7 +203,7 @@ namespace ATech.Ring.DotNet.Cli.Abstractions
             _logger.LogDebug(PhaseStatus.PENDING);
             await StartAsync(ctx, token);
             _logger.LogContextDebug(ctx);
-            _logger.LogDebug(PhaseStatus.OK);
+            _logger.LogInformation(PhaseStatus.OK);
             Sender.Enqueue(RunnableEvent.New<RunnableStarted>(this));
         }
         private async Task<HealthStatus> CheckHealthCoreAsync(TContext ctx, CancellationToken token)