+++ src/source/ATech.Ring.DotNet.Cli/Windows/Tools/GitClone.cs
@@ -41,10 +41,12 @@ namespace ATech.Ring.DotNet.Cli.Windows.Tools
             return Path.IsPathRooted(targetPath) ? targetPath : Path.GetFullPath(targetPath);
         }
 
-        public async Task<ExecutionInfo> CloneOrPullAsync(IFromGit gitCfg, string rootPathOverride = null)
+        public async Task<ExecutionInfo> CloneOrPullAsync(IFromGit gitCfg, bool shallow = false, bool masterOnly = false, string rootPathOverride = null)
         {
             using var _ = Logger.WithScope(gitCfg.SshRepoUrl, Phase.GIT);
-
+            var depthArg = shallow ? "--depth=1" : "";
+            var branchArg = masterOnly ? "--branch=master" : "";
+            var singleBranchArg = masterOnly ? "--single-branch" : "";
             var cloneFullPath = ResolveFullClonePath(gitCfg, rootPathOverride);
 
             async Task<ExecutionInfo> CloneAsync()
@@ -54,7 +56,7 @@ namespace ATech.Ring.DotNet.Cli.Windows.Tools
                 return await this.TryAsync(3, TimeSpan.FromSeconds(10),
                     async t =>
                     {
-                        var result = await t.RunProcessWaitAsync("clone", gitCfg.SshRepoUrl, cloneFullPath);
+                        var result = await t.RunProcessWaitAsync("clone", branchArg, singleBranchArg, depthArg, "--", gitCfg.SshRepoUrl, cloneFullPath);
                         Logger.LogInformation(result.IsSuccess ? PhaseStatus.OK : PhaseStatus.FAILED);
                         return result;
                     });
@@ -69,7 +71,7 @@ namespace ATech.Ring.DotNet.Cli.Windows.Tools
                 return await this.TryAsync(3, TimeSpan.FromSeconds(10),
                     async t =>
                     {
-                        var result = await t.RunProcessWaitAsync("-C", cloneFullPath, "pull");
+                        var result = await t.RunProcessWaitAsync("-C", cloneFullPath, "pull", depthArg);
                         Logger.LogInformation(result.IsSuccess ? PhaseStatus.OK : PhaseStatus.FAILED);
                         return result;
                     });