+++ src/source/ATech.Ring.DotNet.Cli/Windows/Tools/Kubectl.cs
@@ -12,9 +12,26 @@ namespace ATech.Ring.DotNet.Cli.Windows.Tools
         public string[] DefaultArgs { get; set; } = { };
         public ILogger<ITool> Logger { get; }
 
-        public async Task<ExecutionInfo> ApplyKJsonPathAsync(string kustomizeDir, string jsonPath)
+        public async Task<bool> IsValidManifestAsync(string filePath)
         {
-            return await this.RunProcessWaitAsync("kustomize", "build", $"\"{kustomizeDir}\"", "|", "kubectl", "apply", "-o", $"jsonpath=\"{jsonPath}\"", "-f", "-");
+            var result = await this.RunProcessWaitAsync("kubectl", "apply", "--validate=true", "--dry-run=true", "-f", $"\"{filePath}\"");
+            return result.IsSuccess;
+        }
+
+        public async Task<bool> FileExistsAsync(string filePath)
+        {
+            var result = await this.RunProcessWaitAsync($"[ -f \"{filePath}\" ] && echo \"true\" || echo \"false\"");
+            return bool.Parse(result.Output);
+        }
+
+        public async Task<ExecutionInfo> KustomizeBuildAsync(string kustomizeDir, string outputFilePath)
+        {
+            return await this.RunProcessWaitAsync("kustomize", "build", $"\"{kustomizeDir}\"", ">", outputFilePath);
+        }
+
+        public async Task<ExecutionInfo> ApplyJsonPathAsync(string path, string jsonPath)
+        {
+            return await this.RunProcessWaitAsync("kubectl", "apply", "-o", $"jsonpath=\"{jsonPath}\"", "-f", $"\"{path}\"");
         }
 
         public async Task<ExecutionInfo> GetResources(string kind, string nameSpace)
@@ -27,9 +44,9 @@ namespace ATech.Ring.DotNet.Cli.Windows.Tools
             return await this.RunProcessWaitAsync("kubectl", "get", podName, "-o", "jsonpath='{.status.phase}'", "-n", nameSpace);
         }
 
-        public async Task<ExecutionInfo> DeleteKAsync(string kustomizeDir)
+        public async Task<ExecutionInfo> DeleteAsync(string path)
         {
-            return await this.RunProcessWaitAsync("kustomize", "build", $"\"{kustomizeDir}\"", "|", "kubectl", "delete", "-o", "name", "-f", "-");
+            return await this.RunProcessWaitAsync("kubectl", "delete", "--ignore-not-found", "-f", $"\"{path}\"");
         }
     }
 }