+++ src/source/ATech.Ring.DotNet.Cli/Workspace/RunnableContainer.cs
@@ -14,11 +14,11 @@ namespace ATech.Ring.DotNet.Cli.Workspace
         public IRunnable Runnable { get; private set; }
         public Task Task { get; private set; }
 
-        private void Initialise(IRunnableConfig cfg, IRunnable runnable, CancellationToken token)
+        private async Task InitialiseAsync(IRunnable runnable, TimeSpan delay, CancellationToken token)
         {
             _aggregateCts = CancellationTokenSource.CreateLinkedTokenSource(_cts.Token, token);
+            if (delay != TimeSpan.Zero) await Task.Delay(delay, _aggregateCts.Token);
             Runnable = runnable;
-            Task = runnable.RunAsync(cfg, _aggregateCts.Token);
         }
 
         public async Task CancelAsync()
@@ -32,10 +32,11 @@ namespace ATech.Ring.DotNet.Cli.Workspace
         {
             var runnable = factory(cfg);
             var container = new RunnableContainer();
-            if (delay != TimeSpan.Zero) await Task.Delay(delay, token);
-            container.Initialise(cfg, runnable, token);
+            await container.InitialiseAsync(runnable, delay, token);
             return container;
         }
+        
+        public void Start() => Task = Runnable.RunAsync(_aggregateCts.Token);
 
         public void Dispose()
         {